requirements.txt
cycler==0.10.0
h5py==3.2.1
Keras==2.4.3
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
opencv-python==4.5.2.52
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv
pytz==2021.1
PyYAML==5.4.1
scipy==1.6.3
six==1.16.0
imutils
seaborn
tensorflow